feat(desktop): remove Electron, polish Electrobun — window state, View menu, graceful shutdown
- Delete electron/ directory (24 files) — Electron is fully superseded
- Remove 'electron' from root package.json workspaces; add 'electrobun'
- Replace all electron-* justfile recipes with electrobun-* equivalents
- Electrobun shell improvements:
  - Persist window dimensions across launches (windowStore via store.ts)
  - Restore saved width/height on BrowserWindow creation
  - Add View menu: Reload, Zoom In/Out/Reset, Toggle Full Screen
  - Add Window menu: Minimize, Zoom, Bring All to Front (macOS standard)
  - Graceful shutdown: server.stop() on win.on('close') to kill spawned
    Next.js/FastAPI processes and avoid ghost processes after quit
  - Store server ref so shutdown handler can reach it after async init
