--- electron/.gitignore
@@ -1,4 +1,0 @@
-dist/
-dist-app/
-node_modules/
-*.log
