--- electron/package.json
@@ -1,101 +1,0 @@
-{
-	"name": "@pawrrtal/electron",
-	"version": "0.1.0",
-	"private": true,
-	"description": "Pawrrtal desktop shell — wraps the Next.js frontend in an Electron BrowserWindow.",
-	"main": "dist/main.js",
-	"scripts": {
-		"dev": "tsc --watch --preserveWatchOutput",
-		"dev:all": "bun run dev-all.ts",
-		"dev:all:full": "bun run dev-all-full.ts",
-		"build": "tsc",
-		"start": "electron .",
-		"start:dev": "ELECTRON_DEV=1 electron .",
-		"test": "vitest run",
-		"test:watch": "vitest",
-		"dist": "bun run build && electron-builder",
-		"dist:mac": "bun run build && electron-builder --mac",
-		"dist:dir": "bun run build && electron-builder --dir"
-	},
-	"dependencies": {
-		"chokidar": "^5.0.0",
-		"electron-store": "^11.0.2",
-		"wait-on": "^9.0.1"
-	},
-	"devDependencies": {
-		"@types/node": "^22.10.0",
-		"@types/wait-on": "^5.3.4",
-		"@vitest/coverage-v8": "^4.1.5",
-		"electron": "^42.0.1",
-		"electron-builder": "^26.8.1",
-		"typescript": "^5.9.3",
-		"vitest": "^4.1.5"
-	},
-	"build": {
-		"appId": "ai.nexus.desktop",
-		"productName": "Pawrrtal",
-		"directories": {
-			"buildResources": "build",
-			"output": "dist-app"
-		},
-		"files": [
-			"dist/**/*",
-			"package.json"
-		],
-		"extraResources": [
-			{
-				"from": "../frontend/.next/standalone",
-				"to": "frontend",
-				"filter": [
-					"**/*"
-				]
-			},
-			{
-				"from": "../frontend/.next/static",
-				"to": "frontend/.next/static",
-				"filter": [
-					"**/*"
-				]
-			},
-			{
-				"from": "../frontend/public",
-				"to": "frontend/public",
-				"filter": [
-					"**/*"
-				]
-			}
-		],
-		"mac": {
-			"category": "public.app-category.productivity",
-			"target": [
-				{
-					"target": "dmg",
-					"arch": [
-						"arm64",
-						"x64"
-					]
-				},
-				{
-					"target": "zip",
-					"arch": [
-						"arm64",
-						"x64"
-					]
-				}
-			],
-			"hardenedRuntime": false,
-			"gatekeeperAssess": false
-		},
-		"win": {
-			"target": [
-				"nsis"
-			]
-		},
-		"linux": {
-			"target": [
-				"AppImage"
-			],
-			"category": "Office"
-		}
-	}
-}
