--- electron/tsconfig.json
@@ -1,22 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "ES2022",
-		"module": "CommonJS",
-		"moduleResolution": "Node",
-		"lib": ["ES2022", "DOM"],
-		"outDir": "dist",
-		"rootDir": "src",
-		"strict": true,
-		"noUncheckedIndexedAccess": true,
-		"noImplicitReturns": true,
-		"noFallthroughCasesInSwitch": true,
-		"esModuleInterop": true,
-		"forceConsistentCasingInFileNames": true,
-		"skipLibCheck": true,
-		"sourceMap": true,
-		"declaration": false,
-		"resolveJsonModule": true,
-		"types": ["node"]
-	},
-	"include": ["src/**/*"]
-}
