+++ package.json
@@ -26,7 +26,7 @@
 	},
 	"workspaces": [
 		"frontend",
-		"electron",
+		"electrobun",
 		"frontend/lib/react-dropdown"
 	],
 	"overrides": {
